requirements.txt
# Libraries
langchain
langchain-core
langchain-community
langchain-together
pydantic
typing-extensions
requests
beautifulsoup4
python-dotenv
google-adk
google-genai
tavily-python
faiss-cpu
sentence-transformers
torch
torchvision
torchaudio
